kre-py/requirements.txt
asyncio-nats-client==0.10.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
future==0.18.2
gensim==3.4.0
idna==2.10
influxdb-client==1.9.0
joblib==0.17.0
nltk==3.5
numpy==1.18.4
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.1
protobuf==3.13.0
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.4
regex==2020.10.28
requests==2.24.0
Rx==3.1.1
scikit-learn==0.23.1
scipy==1.5.4
six==1.15.0
smart-open==3.0.0
threadpoolctl==2.1.0
torch==1.5.1
torchvision==0.6.1
tqdm==4.51.0
urllib3==1.25.11
